split collation logic and tests in 2.12 and 2.13 folders
